chore: refactor merge utils
