+++ .changeset/honest-feet-design.md
@@ -1,0 +1,9 @@
+---
+'@pandacss/config': patch
+---
+
+Create separate entrypoint for merge configs
+
+```ts
+import { mergeConfigs } from '@pandacss/config/merge'
+```
