+++ packages/config/merge.d.ts
@@ -1,0 +1,1 @@
+export * from './src/merge.ts'
