+++ packages/config/src/index.ts
@@ -1,4 +1,4 @@
 export { findConfigFile } from './find-config'
 export { getConfigDependencies, type GetConfigDependenciesTsOptions, type GetDepsOptions } from './get-mod-deps'
 export { loadConfigFile, resolveConfigFile, bundleConfigFile } from './load-config'
-export { getResolvedConfig, mergeConfigs } from './merge-config'
+export { getResolvedConfig } from './get-resolved-config'
