+++ packages/config/src/merge.ts
@@ -1,6 +1,5 @@
 import type { Config } from '@pandacss/types'
 import { mergeAndConcat } from 'merge-anything'
-import { bundle } from './bundle'
 import { assign, mergeWith } from './utils'
 
 type Extendable<T> = T & { extend?: T }
@@ -80,26 +79,4 @@ export function mergeConfigs(configs: ExtendableConfig[]) {
   )
 
   return compact(mergedResult)
-}
-
-/**
- * Recursively merge all presets into a single config
- */
-export async function getResolvedConfig(config: ExtendableConfig, cwd: string) {
-  const presets = config.presets ?? []
-
-  const configs: ExtendableConfig[] = [config]
-  while (presets.length > 0) {
-    const preset = await presets.shift()!
-    if (typeof preset === 'string') {
-      const presetModule = await bundle(preset, cwd)
-      configs.push(await presetModule.config)
-      presets.unshift(...((await presetModule.config.presets) ?? []))
-    } else {
-      configs.push(preset)
-      presets.unshift(...(preset.presets ?? []))
-    }
-  }
-
-  return mergeConfigs(configs) as Config
 }
