+++ packages/shared/package.json
