+++ packages/studio/styled-system/helpers.mjs
@@ -73,7 +73,6 @@ function mergeProps(...sources) {
 }
 
 // src/walk-object.ts
-var isNotNullish = (element) => element != null
 function walkObject(target, predicate, options = {}) {
   const { stop, getKey } = options
   function inner(value, path = []) {
@@ -85,10 +84,7 @@ function walkObject(target, predicate, options = {}) {
         if (stop?.(value, childPath)) {
           return predicate(value, path)
         }
-        const next = inner(child, childPath)
-        if (isNotNullish(next)) {
-          result[key] = next
-        }
+        result[key] = inner(child, childPath)
       }
       return result
     }
