+++ playground/package.json
@@ -24,6 +24,7 @@
     "@pandacss/parser": "workspace:*",
     "@pandacss/preset-base": "workspace:*",
     "@pandacss/preset-panda": "workspace:*",
+    "@pandacss/config": "workspace:*",
     "@prisma/client": "4.16.1",
     "@textea/json-viewer": "3.1.1",
     "hookable": "5.5.3",
